refactor(files): extract folder data hook
